requirements.txt
python-telegram-bot==20.7
httpx==0.26.0
gunicorn==20.1.0
Flask==2.2.2